Moved App files to component folder
